+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.3]
+## [0.2.4] - 2018-12-31
+### Added
+- Ability to update the video URL via hitting an Enter key.
+- Error message for the server name / passphrase panel.
+
+### Fixed
+- Patched the ability to create an empty room name as it will conflict the
+frontend's ability to update itself.
+
+## [0.2.3] - 2018-12-31
 ### Added
 - A changelog
 - Ability for the video player to dynamically resize to the user's screen
@@ -17,10 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 room.
 - Fixed a bug that did not reset the video player after the user left the room.
 
-## [0.2.1]
+## [0.2.1] - 2018-12-31
 ### Added
 - Docker support
 
-## [0.2.0]
+## [0.2.0] - 2018-12-30
 ### Added
 - First release of the working website